+++ src/internal/connector/graph_connector_onedrive_test.go
@@ -286,7 +286,7 @@ type itemData struct {
 	perms permData
 }
 
-type onedriveColInfo struct {
+type driveColInfo struct {
 	pathElements []string
 	perms        permData
 	files        []itemData
@@ -296,7 +296,7 @@ type onedriveColInfo struct {
 func testDataForInfo(
 	t *testing.T,
 	service path.ServiceType,
-	cols []onedriveColInfo,
+	cols []driveColInfo,
 	backupVersion int,
 ) []colInfo {
 	var res []colInfo
@@ -431,11 +431,6 @@ func (si suiteInfoImpl) Resource() Resource {
 // SharePoint shares most of its libraries implementation with OneDrive so we
 // only test simple things here and leave the more extensive testing to
 // OneDrive.
-//
-// TODO(ashmrtn): SharePoint doesn't have permissions backup/restore enabled
-// right now. Adjust the tests here when that is enabled so we have at least
-// basic assurances that it's doing the right thing. We can leave the more
-// extensive permissions tests to OneDrive as well.
 
 type GraphConnectorSharePointIntegrationSuite struct {
 	tester.Suite
@@ -484,6 +479,23 @@ func (suite *GraphConnectorSharePointIntegrationSuite) SetupSuite() {
 
 func (suite *GraphConnectorSharePointIntegrationSuite) TestRestoreAndBackup_MultipleFilesAndFolders_NoPermissions() {
 	testRestoreAndBackupMultipleFilesAndFoldersNoPermissions(suite, version.Backup)
+}
+
+func (suite *GraphConnectorSharePointIntegrationSuite) TestPermissionsRestoreAndBackup() {
+	testPermissionsRestoreAndBackup(suite, version.Backup)
+}
+
+func (suite *GraphConnectorSharePointIntegrationSuite) TestPermissionsBackupAndNoRestore() {
+	testPermissionsBackupAndNoRestore(suite, version.Backup)
+}
+
+func (suite *GraphConnectorSharePointIntegrationSuite) TestPermissionsInheritanceRestoreAndBackup() {
+	testPermissionsInheritanceRestoreAndBackup(suite, version.Backup)
+}
+
+func (suite *GraphConnectorSharePointIntegrationSuite) TestRestoreFolderNamedFolderRegression() {
+	// No reason why it couldn't work with previous versions, but this is when it got introduced.
+	testRestoreFolderNamedFolderRegression(suite, version.All8MigrateUserPNToID)
 }
 
 // ---------------------------------------------------------------------------
@@ -663,7 +675,7 @@ func testRestoreAndBackupMultipleFilesAndFoldersNoPermissions(
 		folderBName,
 	}
 
-	cols := []onedriveColInfo{
+	cols := []driveColInfo{
 		{
 			pathElements: rootPath,
 			files: []itemData{
@@ -807,7 +819,7 @@ func testPermissionsRestoreAndBackup(suite oneDriveSuite, startVersion int) {
 		folderCName,
 	}
 
-	cols := []onedriveColInfo{
+	cols := []driveColInfo{
 		{
 			pathElements: rootPath,
 			files: []itemData{
@@ -939,9 +951,10 @@ func testPermissionsRestoreAndBackup(suite oneDriveSuite, startVersion int) {
 	}
 
 	expected := testDataForInfo(suite.T(), suite.BackupService(), cols, version.Backup)
+	bss := suite.BackupService().String()
 
 	for vn := startVersion; vn <= version.Backup; vn++ {
-		suite.Run(fmt.Sprintf("Version%d", vn), func() {
+		suite.Run(fmt.Sprintf("%s-Version%d", bss, vn), func() {
 			t := suite.T()
 			// Ideally this can always be true or false and still
 			// work, but limiting older versions to use emails so as
@@ -984,7 +997,7 @@ func testPermissionsBackupAndNoRestore(suite oneDriveSuite, startVersion int) {
 		suite.Service(),
 		suite.BackupResourceOwner())
 
-	inputCols := []onedriveColInfo{
+	inputCols := []driveColInfo{
 		{
 			pathElements: []string{
 				odConsts.DrivesPathDir,
@@ -1005,7 +1018,7 @@ func testPermissionsBackupAndNoRestore(suite oneDriveSuite, startVersion int) {
 		},
 	}
 
-	expectedCols := []onedriveColInfo{
+	expectedCols := []driveColInfo{
 		{
 			pathElements: []string{
 				odConsts.DrivesPathDir,
@@ -1023,9 +1036,10 @@ func testPermissionsBackupAndNoRestore(suite oneDriveSuite, startVersion int) {
 	}
 
 	expected := testDataForInfo(suite.T(), suite.BackupService(), expectedCols, version.Backup)
+	bss := suite.BackupService().String()
 
 	for vn := startVersion; vn <= version.Backup; vn++ {
-		suite.Run(fmt.Sprintf("Version%d", vn), func() {
+		suite.Run(fmt.Sprintf("%s-Version%d", bss, vn), func() {
 			t := suite.T()
 			input := testDataForInfo(t, suite.BackupService(), inputCols, vn)
 
@@ -1150,7 +1164,7 @@ func testPermissionsInheritanceRestoreAndBackup(suite oneDriveSuite, startVersio
 	// 	   - inherted-permission-file
 	//     - empty-permission-file (empty/empty might have interesting behavior)
 
-	cols := []onedriveColInfo{
+	cols := []driveColInfo{
 		{
 			pathElements: rootPath,
 			files:        []itemData{},
@@ -1199,9 +1213,10 @@ func testPermissionsInheritanceRestoreAndBackup(suite oneDriveSuite, startVersio
 	}
 
 	expected := testDataForInfo(suite.T(), suite.BackupService(), cols, version.Backup)
+	bss := suite.BackupService().String()
 
 	for vn := startVersion; vn <= version.Backup; vn++ {
-		suite.Run(fmt.Sprintf("Version%d", vn), func() {
+		suite.Run(fmt.Sprintf("%s-Version%d", bss, vn), func() {
 			t := suite.T()
 			// Ideally this can always be true or false and still
 			// work, but limiting older versions to use emails so as
@@ -1264,7 +1279,7 @@ func testRestoreFolderNamedFolderRegression(
 		folderBName,
 	}
 
-	cols := []onedriveColInfo{
+	cols := []driveColInfo{
 		{
 			pathElements: rootPath,
 			files: []itemData{
@@ -1313,9 +1328,10 @@ func testRestoreFolderNamedFolderRegression(
 	}
 
 	expected := testDataForInfo(suite.T(), suite.BackupService(), cols, version.Backup)
+	bss := suite.BackupService().String()
 
 	for vn := startVersion; vn <= version.Backup; vn++ {
-		suite.Run(fmt.Sprintf("Version%d", vn), func() {
+		suite.Run(fmt.Sprintf("%s-Version%d", bss, vn), func() {
 			t := suite.T()
 			input := testDataForInfo(t, suite.BackupService(), cols, vn)