+++ src/internal/connector/graph_connector_test.go
@@ -430,8 +430,7 @@ func getCollectionsAndExpected(
 			owner,
 			config.dest,
 			testCollections,
-			backupVersion,
-		)
+			backupVersion)
 
 		collections = append(collections, ownerCollections...)
 		totalItems += numItems
@@ -550,8 +549,7 @@ func runBackupAndCompare(
 		totalKopiaItems,
 		expectedData,
 		dcs,
-		config.dest,
-		config.opts.RestorePermissions)
+		config)
 
 	status := backupGC.Wait()
 
@@ -1125,17 +1123,15 @@ func (suite *GraphConnectorIntegrationSuite) TestMultiFolderBackupDifferentNames
 
 			t.Log("Backup enumeration complete")
 
+			ci := configInfo{
+				opts: control.Options{RestorePermissions: true},
+				// Alright to be empty, needed for OneDrive.
+				dest: control.RestoreDestination{},
+			}
+
 			// Pull the data prior to waiting for the status as otherwise it will
 			// deadlock.
-			skipped := checkCollections(
-				t,
-				ctx,
-				allItems,
-				allExpectedData,
-				dcs,
-				// Alright to be empty, needed for OneDrive.
-				control.RestoreDestination{},
-				true)
+			skipped := checkCollections(t, ctx, allItems, allExpectedData, dcs, ci)
 
 			status := backupGC.Wait()
 			assert.Equal(t, allItems+skipped, status.Objects, "status.Objects")